Java-Frontend: Reorganize java frontend code
